feat: Transitions funtional with shared features.
still could use work with return animation looking chunking when it shrinks to a list item.

If you select a list item and swipe to another in the detail screen and than return to the list view, the return animation puts the item in the place where you originally clicked rather than the spot where the item should be.